(fix) Adding links for the device selector.
Refs #25281
