+++ core-web/apps/dotcms-ui/src/app/portlets/dot-edit-page/seo/components/dot-device-selector-seo/dot-device-selector-seo.component.html
@@ -9,7 +9,12 @@
                 <h3 class="dot-device-selector__header-title">
                     {{ 'editpage.device.selector.title' | dm }}
                 </h3>
-                <p-button icon="pi pi-plus" styleClass="p-button-rounded p-button-text"></p-button>
+                <a [routerLink]="linkToAddDevice" data-testId="dot-device-link-add">
+                    <p-button
+                        icon="pi pi-plus"
+                        styleClass="p-button-rounded p-button-text"
+                    ></p-button>
+                </a>
             </div>
             <div class="dot-device-selector__grid">
                 <ul class="device-list">
@@ -81,7 +86,14 @@
             </div>
         </div>
         <div class="dot-device-selector__link--container">
-            <a class="dot-device-selector__link"> {{ 'editpage.device.selector.new.tab' | dm }} </a>
+            <a
+                class="dot-device-selector__link"
+                [href]="previewUrl"
+                data-testId="dot-device-selector-link"
+                target="_blank"
+            >
+                {{ 'editpage.device.selector.new.tab' | dm }}
+            </a>
             <i class="pi pi-external-link"></i>
         </div>
     </div>
