+++ core-web/apps/dotcms-ui/src/app/portlets/dot-edit-page/seo/components/dot-device-selector-seo/dot-device-selector-seo.component.spec.ts
@@ -5,6 +5,7 @@ import { CUSTOM_ELEMENTS_SCHEMA, Component, DebugElement } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { RouterTestingModule } from '@angular/router/testing';
 
 import { OverlayPanelModule } from 'primeng/overlaypanel';
 
@@ -20,10 +21,13 @@ import { DotDeviceSelectorSeoComponent } from './dot-device-selector-seo.compone
 
 @Component({
     selector: 'dot-test-host-component',
-    template: `<button (click)="op.openMenu($event)" type="text"></button>
-        <dot-device-selector-seo #op></dot-device-selector-seo> `
+    template: `<button (click)="op.openMenu($event)" type="text">Open</button>
+        <dot-device-selector-seo #op [apiLink]="apiLink"></dot-device-selector-seo> `
 })
-class TestHostComponent {}
+class TestHostComponent {
+    apiLink = 'api/v1/page/render/an/url/test?language_id=1';
+    linkToAddDevice = '/c/c_Devices';
+}
 
 describe('DotDeviceSelectorSeoComponent', () => {
     let fixtureHost: ComponentFixture<TestHostComponent>;
@@ -51,7 +55,8 @@ describe('DotDeviceSelectorSeoComponent', () => {
                 DotDeviceSelectorSeoComponent,
                 HttpClientTestingModule,
                 OverlayPanelModule,
-                BrowserAnimationsModule
+                BrowserAnimationsModule,
+                RouterTestingModule
             ],
             providers: [
                 {
@@ -80,12 +85,11 @@ describe('DotDeviceSelectorSeoComponent', () => {
         spyOn(component, 'getOptions').and.returnValue(of(mockDotDevices));
 
         fixtureHost.detectChanges();
+        const buttonEl = fixtureHost.debugElement.query(By.css('button')).nativeElement;
+        buttonEl.click();
     });
 
     it('should emit selected device on change', async () => {
-        const buttonEl = fixtureHost.debugElement.query(By.css('button')).nativeElement;
-        buttonEl.click();
-
         await fixtureHost.whenStable();
         fixtureHost.detectChanges();
         spyOn(component.selected, 'emit');
@@ -101,9 +105,6 @@ describe('DotDeviceSelectorSeoComponent', () => {
     });
 
     it('should set user devices', async () => {
-        const buttonEl = fixtureHost.debugElement.query(By.css('button')).nativeElement;
-        buttonEl.click();
-
         await fixtureHost.whenStable();
         fixtureHost.detectChanges();
 
@@ -126,14 +127,27 @@ describe('DotDeviceSelectorSeoComponent', () => {
     });
 
     it('should close the overlayPanel', () => {
-        const buttonEl = fixtureHost.debugElement.query(By.css('button')).nativeElement;
-        buttonEl.click();
-
         const devicesSelector: DebugElement = de.query(
             By.css('[data-testId="dot-devices-selector"]')
         );
-
-        fixtureHost.detectChanges();
         expect(devicesSelector).toBeNull();
+    });
+
+    it('should have link to open in a new tab', () => {
+        fixtureHost.detectChanges();
+
+        const addContent: DebugElement = de.query(
+            By.css('[data-testId="dot-device-selector-link"]')
+        );
+        expect(addContent.nativeElement.href).toContain(
+            '/an/url/test?language_id=1&disabledNavigateMode=true'
+        );
+    });
+
+    it('should have a link to add device', () => {
+        fixtureHost.detectChanges();
+
+        const link = de.query(By.css('[data-testId="dot-device-link-add"]'));
+        expect(link.properties.href).toContain('/c/c_Devices');
     });
 });
