+++ core-web/apps/dotcms-ui/src/app/portlets/dot-edit-page/seo/components/dot-edit-page-state-controller-seo/dot-edit-page-state-controller-seo.component.html
@@ -1,6 +1,7 @@
 <dot-device-selector-seo
     #deviceSelector
     [pTooltip]="pageState.viewAs.device?.name || ('editpage.viewas.default.device' | dm)"
+    [apiLink]="apiLink"
     (selected)="changeDeviceHandler($event)"
     appendTo="body"
     tooltipPosition="bottom"
