+++ core-web/apps/dotcms-ui/src/app/portlets/dot-edit-page/seo/components/dot-edit-page-toolbar-seo/dot-edit-page-toolbar-seo.component.spec.ts
@@ -279,9 +279,6 @@ describe('DotEditPageToolbarSeoComponent', () => {
             const dotEditPageInfo = de.query(By.css('dot-edit-page-info-seo')).componentInstance;
             expect(dotEditPageInfo.title).toBe('A title');
             expect(dotEditPageInfo.url).toBe('/an/url/test');
-            expect(dotEditPageInfo.innerApiLink).toBe(
-                'api/v1/page/render/an/url/test?language_id=1'
-            );
         });
     });
